requirements.txt
fastapi==0.115.6
pydantic==2.10.4
pymongo==4.10.1
python_bcrypt==0.3.2
starlette==0.41.0  
uvicorn==0.22.0
python-dotenv==1.0.0
itsdangerous==2.1.2